exsubs_pkg/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "exsubs"
version = "0.1.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "A tool to extract and translate subtitles from MKV files."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "filelock",
    "python-dotenv",
    "rich",
    "subtitle_filter",
]

[project.scripts]
exsubs = "exsubs.cli:main"
